feat: redux typescript 문서 반영
